v1.0
When run, this program, will ask you a couple questions on how to handle your images.  All inputs are either "y" or "n" except for the target address that you are searching for images.  

Create new files Locally: It may not be possible to directly edit, save and delete the file from the server. It may also be faster to do the conversion locally, compare and delete, then push back the compressed copy. That is also the safest. 


Delete on the fly: Choosing to delete here only applies to the source used for the downscaling. If you have to create the compressed files on your computer first, enter Y, as it will not touch the Originals: only the newly created copy. If you are working in the server directory without making a local copy first: THIS WILL DELETE THE ORIGINAL without further prompting as the compressed file is made. please test that files are reliably made remotely BEFORE attempting deletion on the fly. 

Comparison delete: Adding in safety, this gets the list of jpegs again from the remote target and if there is a matching compressed file, it is deleted.